requirements.txt
toml
requests
python-socketio